refactor(parser): split the two modules the IfcMapConversionScaled fix pushed over budget
Neither ratchet budget is raised; both files shrink. Pure moves — no
expression and no control flow changed.

georef-extractor.ts (675 > 653): the local/world arithmetic
(computeTransformMatrix, transformToWorld, transformToLocal,
getCoordinateSystemDescription) moves to georef-transform.ts, which reads
MapConversion / GeoreferenceInfo as a TYPE-only import, so no runtime
cycle is created. The three public functions are re-exported from
georef-extractor.ts; computeTransformMatrix stays module-private as
before. 675 to 564.

on-demand-extractors.ts (1175 > 1168): extractGeoreferencingOnDemand and
its compute half move to on-demand-georeferencing.ts, and the
oncePerStore memo they share with future extract*OnDemand callers moves
to on-demand-cache.ts so neither new module imports back into the old one.
Entry point re-exported. 1175 to 1037.

Allowlist rows LOWERED, none raised: georef-extractor.ts 653 to 564,
on-demand-extractors.ts 1168 to 1037. ALLOWLIST_DIGEST re-pinned in the
same commit. The three new modules are 133 / 144 / 23 lines and need no
allowlist entry.
