+++ scripts/module-size-allowlist.txt
@@ -276,9 +276,9 @@
   1237 packages/parser/src/columnar-parser.ts
    480 packages/parser/src/compact-entity-index.ts
    530 packages/parser/src/data-store-transport.ts
-   653 packages/parser/src/georef-extractor.ts
+   564 packages/parser/src/georef-extractor.ts
    827 packages/parser/src/material-resolver.ts
-  1168 packages/parser/src/on-demand-extractors.ts
+  1037 packages/parser/src/on-demand-extractors.ts
    523 packages/parser/src/project-units.ts
    594 packages/parser/src/schedule-extractor.ts
    494 packages/parser/src/source-bytes.ts
